Add sqlite3 db script and support

+++ benoic.mariadb.sql
@@ -40,7 +40,7 @@ CREATE TABLE `b_element` (
   `be_enabled` tinyint(1) DEFAULT 1,
   `be_options` BLOB,
   `be_monitored` tinyint(1) DEFAULT 0,
-  `be_monitored_every` tinyint(1) DEFAULT 0,
+  `be_monitored_every` int(11) DEFAULT 0,
   `be_monitored_next` timestamp,
   CONSTRAINT `device_ibfk_1` FOREIGN KEY (`bd_name`) REFERENCES `b_device` (`bd_name`) ON DELETE CASCADE
 );
